# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Order Fulfillment and Logistics Hub API
  version: 0.1.0
  description: >-
    Contract for BUSINESS administration, trusted internal intake, and isolated
    DemoScenario operations. JSON field names use snake_case; identifiers and
    NUMERIC(18,3) quantities are strings.
servers:
  - url: http://localhost:8080
    description: Local Docker Compose
...
security:
  - basicAuth: []
paths:
...
  /api/v1/order-drafts:
    get:
      tags: [OrderDrafts]
      operationId: listOrderDrafts
      description: Whitelisted enterprise WeChat order drafts; model-private and unknown fields are never returned.
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderDraftStatus' }
        - name: submission_id
          in: query
          schema: { $ref: '#/components/schemas/Identifier' }
      responses:
        '200':
          description: Newest drafts first
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDraftPage' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/order-drafts/{draft_id}:
    parameters:
      - $ref: '#/components/parameters/OrderDraftId'
    get:
      tags: [OrderDrafts]
      operationId: getOrderDraft
      responses:
        '200':
          description: Whitelisted draft, candidate, line, and open ReviewCase version data
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDraftDetail' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/order-drafts/{draft_id}/confirm:
    parameters:
      - $ref: '#/components/parameters/OrderDraftId'
    post:
      tags: [OrderDrafts]
      operationId: confirmOrderDraft
      description: >-
        Idempotently confirm the visible draft and ReviewCase versions and create one WECOM CanonicalOrder.
        The backend revalidates the gateway Basic Auth principal and requires it to match X-Operator;
        the header alone is never an authenticated identity.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ConfirmOrderDraftCommand' }
      responses:
        '200':
          description: Confirmed draft; an identical idempotency replay returns the same representation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDraftDetail' }
        '403':
          description: Authenticated gateway principal is missing or does not match X-Operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
  /api/v1/order-drafts/{draft_id}/reject:
    parameters:
      - $ref: '#/components/parameters/OrderDraftId'
    post:
      tags: [OrderDrafts]
      operationId: rejectOrderDraft
      description: >-
        Idempotently reject the visible draft and dismiss its ReviewCase with an explicit reason.
        The backend revalidates the gateway Basic Auth principal and requires it to match X-Operator;
        the header alone is never an authenticated identity.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Operator'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RejectOrderDraftCommand' }
      responses:
        '200':
          description: Rejected draft; an identical idempotency replay returns the same representation
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDraftDetail' }
        '403':
          description: Authenticated gateway principal is missing or does not match X-Operator
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ApiError' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
  /api/v1/orders/{order_id}:
    parameters:
      - $ref: '#/components/parameters/OrderId'
    get:
      tags: [Orders]
      operationId: getOrder
      responses:
        '200':
          description: Aggregated BUSINESS order detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderDetail' }
        default: { $ref: '#/components/responses/ErrorResponse' }
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
      description: Nginx authenticates management, internal, and demo operations; explicitly public connector callbacks use their own signatures.
  headers:
    RequestId:
      description: Request correlation identifier
      schema: { type: string }
  responses:
    ErrorResponse:
      description: Unified business or technical error
      headers:
        X-Request-Id: { $ref: '#/components/headers/RequestId' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 20 }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 8, maxLength: 255 }
    Operator:
      name: X-Operator
      in: header
      required: true
      description: Required at the backend seam; the public gateway overwrites this value with the authenticated principal.
      schema: { type: string, minLength: 1, maxLength: 128 }
    OrderId: { name: order_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
    OrderDraftId: { name: draft_id, in: path, required: true, schema: { $ref: '#/components/schemas/Identifier' } }
...
  schemas:
    Identifier:
      type: string
      pattern: '^[1-9][0-9]*$'
      example: '42'
    DecimalQuantity:
      type: string
      pattern: '^(0|[1-9][0-9]*)(\.[0-9]{1,3})?$'
      example: '6.000'
    PositiveDecimalQuantity:
      type: string
      pattern: '^(?!0(?:\.0{1,3})?$)(0|[1-9][0-9]*)(\.[0-9]{1,3})?$'
      example: '6.000'
...
    SourceChannel:
      type: string
      enum: [CAISHIXIAN, JUFUBAO, FEIXIANG, WECOM]
...
    OrderStatus:
      type: string
      enum: [RECEIVED, VALIDATED, SKU_MAPPED, FULFILLING, SHIPPED, SYNCED, CLOSED, NEED_REVIEW, OUT_OF_STOCK, PROCUREMENT_PENDING, FULFILLMENT_EXCEPTION, SYNC_FAILED, CANCELLED]
    ProcessingStage:
      type: string
      enum: [NEED_REVIEW, READY_TO_EXPORT, PROCUREMENT_IN_PROGRESS, WAITING_PROVIDER, TRACKING_RECEIVED, RETURN_FILE_READY, COMPLETED, EXCEPTION]
    ProcessingHealth:
      type: string
      enum: [BLUE, YELLOW, RED, GREEN]
      description: Derived display projection; never accepted by a write command.
...
    PageMeta:
      type: object
      required: [page, size, total_elements, total_pages]
      properties:
        page: { type: integer, minimum: 0 }
        size: { type: integer, minimum: 1 }
        total_elements: { type: integer, format: int64, minimum: 0 }
        total_pages: { type: integer, minimum: 0 }
    FieldError:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
    ApiError:
      type: object
      required: [business_code, message, http_status, request_id, trace_id]
      properties:
        business_code: { type: string, maxLength: 64 }
        message: { type: string }
        http_status: { type: integer, minimum: 400, maximum: 599 }
        request_id: { type: string }
        trace_id: { type: string }
        field_errors:
          type: array
          items: { $ref: '#/components/schemas/FieldError' }
        details:
          type: object
          additionalProperties: true
...
    OrderDraftStatus:
      type: string
      enum: [OPEN, CONFIRMED, REJECTED]
    OrderDraftCustomerCandidate:
      type: object
      additionalProperties: false
      required: [customer_id, customer_code, customer_name, matched_by]
      properties:
        customer_id: { $ref: '#/components/schemas/Identifier' }
        customer_code: { type: string }
        customer_name: { type: string }
        matched_by: { type: string, enum: [deterministic-mapping] }
    OrderDraftSkuCandidate:
      type: object
      additionalProperties: false
      required: [sku_id, sku_code, product_name, specification, unit, provider_id, source_sku_ref]
      properties:
        sku_id: { $ref: '#/components/schemas/Identifier' }
        sku_code: { type: string }
        product_name: { type: string }
        specification: { type: string }
        unit: { type: string }
        provider_id: { $ref: '#/components/schemas/Identifier' }
        source_sku_ref: { type: string }
        quantity_multiplier:
          allOf:
            - $ref: '#/components/schemas/DecimalQuantity'
          nullable: true
    OrderDraftLine:
      type: object
      additionalProperties: false
      required: [id, line_no, sku_id, sku_code, sku_candidates, product_name_raw, spec_raw, unit_raw, quantity]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        line_no: { type: integer, minimum: 1 }
        sku_id:
          allOf:
            - $ref: '#/components/schemas/Identifier'
          nullable: true
        sku_code: { type: string, nullable: true }
        sku_candidates:
          type: array
          items: { $ref: '#/components/schemas/OrderDraftSkuCandidate' }
        product_name_raw: { type: string, nullable: true }
        spec_raw: { type: string, nullable: true }
        unit_raw: { type: string, nullable: true }
        quantity:
          allOf:
            - $ref: '#/components/schemas/DecimalQuantity'
          nullable: true
    OrderDraftDetail:
      type: object
      additionalProperties: false
      required:
        - id
        - draft_no
        - source_order_no
        - submission_id
        - status
        - revision
        - customer_id
        - customer_code
        - customer_name
        - customer_candidates
        - customer_name_raw
        - receiver_name
        - receiver_phone
        - receiver_address
        - settlement_method
        - missing_fields
        - lines
        - review_case_id
        - review_case_version
        - confirmed_order_id
        - confirmed_by
        - confirmed_at
        - created_at
        - updated_at
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        draft_no: { type: string }
        source_order_no: { type: string }
        submission_id: { $ref: '#/components/schemas/Identifier' }
        status: { $ref: '#/components/schemas/OrderDraftStatus' }
        revision: { type: integer, format: int64, minimum: 0 }
        customer_id:
          allOf:
            - $ref: '#/components/schemas/Identifier'
          nullable: true
        customer_code: { type: string, nullable: true }
        customer_name: { type: string, nullable: true }
        customer_candidates:
          type: array
          items: { $ref: '#/components/schemas/OrderDraftCustomerCandidate' }
        customer_name_raw: { type: string, nullable: true }
        receiver_name: { type: string, nullable: true }
        receiver_phone: { type: string, nullable: true }
        receiver_address: { type: string, nullable: true }
        settlement_method: { type: string, nullable: true }
        missing_fields:
          type: array
          items: { type: string }
        lines:
          type: array
          items: { $ref: '#/components/schemas/OrderDraftLine' }
        review_case_id:
          allOf:
            - $ref: '#/components/schemas/Identifier'
          nullable: true
        review_case_version:
          type: integer
          format: int64
          minimum: 0
          nullable: true
        confirmed_order_id:
          allOf:
            - $ref: '#/components/schemas/Identifier'
          nullable: true
        confirmed_by: { type: string, nullable: true }
        confirmed_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    OrderDraftPage:
      allOf:
        - $ref: '#/components/schemas/PageMeta'
        - type: object
          required: [items]
          properties:
            items:
              type: array
              items: { $ref: '#/components/schemas/OrderDraftDetail' }
    OrderDraftCustomerChoice:
      type: object
      additionalProperties: false
      required: [customer_id]
      properties:
        customer_id: { $ref: '#/components/schemas/Identifier' }
    ConfirmOrderDraftItem:
      type: object
      additionalProperties: false
      required: [line_no, sku_id, quantity]
      properties:
        line_no: { type: integer, minimum: 1 }
        sku_id: { $ref: '#/components/schemas/Identifier' }
        quantity: { $ref: '#/components/schemas/PositiveDecimalQuantity' }
    ConfirmOrderDraftCommand:
      type: object
      additionalProperties: false
      required: [expected_revision, expected_case_version, customer, receiver, settlement, items]
      properties:
        expected_revision: { type: integer, format: int64, minimum: 0 }
        expected_case_version: { type: integer, format: int64, minimum: 0 }
        customer: { $ref: '#/components/schemas/OrderDraftCustomerChoice' }
        receiver: { $ref: '#/components/schemas/Receiver' }
        settlement: { $ref: '#/components/schemas/Settlement' }
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/ConfirmOrderDraftItem' }
        remark: { type: string, maxLength: 2000 }
    RejectOrderDraftCommand:
      type: object
      additionalProperties: false
      required: [expected_revision, expected_case_version, reason]
      properties:
        expected_revision: { type: integer, format: int64, minimum: 0 }
        expected_case_version: { type: integer, format: int64, minimum: 0 }
        reason: { type: string, minLength: 1, maxLength: 2000 }
...
    Receiver:
      type: object
      required: [name, phone, address]
      properties:
        name: { type: string, maxLength: 128 }
        phone: { type: string, maxLength: 64 }
        province: { type: string, maxLength: 64 }
        city: { type: string, maxLength: 64 }
        district: { type: string, maxLength: 64 }
        town: { type: string, maxLength: 64 }
        address: { type: string, maxLength: 1000 }
    Settlement:
      type: object
      required: [method, settlement_time]
      properties:
        method: { type: string, enum: [MONTHLY, IMMEDIATE, CREDIT_TERM, PREPAID, COD, OTHER] }
        settlement_time: { type: string, format: date-time }
...
    OrderLine:
      type: object
      required: [id, line_no, line_type, product_name, specification, unit, requested_quantity, processing_stage]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        line_no: { type: integer, minimum: 1 }
        line_type: { type: string, enum: [SINGLE, CUSTOM_BUNDLE] }
        sku_id: { $ref: '#/components/schemas/Identifier' }
        sku_code: { type: string }
        provider_id: { $ref: '#/components/schemas/Identifier' }
        product_name: { type: string }
        specification: { type: string }
        unit: { type: string }
        source_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        mapping_multiplier: { $ref: '#/components/schemas/DecimalQuantity' }
        requested_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
        processing_stage: { $ref: '#/components/schemas/ProcessingStage' }
        exception_code: { type: string }
        components:
          type: array
          items: { $ref: '#/components/schemas/OrderLineComponent' }
    OrderLineComponent:
      type: object
      required: [id, sku_id, product_name, specification, unit, quantity_per_bundle, total_quantity]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        sku_id: { $ref: '#/components/schemas/Identifier' }
        product_name: { type: string }
        specification: { type: string }
        unit: { type: string }
        quantity_per_bundle: { $ref: '#/components/schemas/DecimalQuantity' }
        total_quantity: { $ref: '#/components/schemas/DecimalQuantity' }
    OrderSummary:
      type: object
      required: [id, order_no, source_channel, order_status, processing_stage, processing_health, completed_count, total_count, created_at, version]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        order_no: { type: string }
        source_channel: { $ref: '#/components/schemas/SourceChannel' }
        source_ref: { type: string }
        customer_id: { $ref: '#/components/schemas/Identifier' }
        customer_name: { type: string }
        receiver_name: { type: string }
        order_status: { $ref: '#/components/schemas/OrderStatus' }
        processing_stage: { $ref: '#/components/schemas/ProcessingStage' }
        processing_health: { $ref: '#/components/schemas/ProcessingHealth' }
        completed_count: { type: integer, minimum: 0 }
        total_count: { type: integer, minimum: 1 }
        attention_reason: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        version: { type: integer, format: int64, minimum: 0 }
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          required: [receiver, settlement, lines, review_cases]
          properties:
            receiver: { $ref: '#/components/schemas/Receiver' }
            settlement: { $ref: '#/components/schemas/Settlement' }
            remark: { type: string }
            lines:
              type: array
              items: { $ref: '#/components/schemas/OrderLine' }
            review_cases:
              type: array
              items: { $ref: '#/components/schemas/ReviewCase' }
...
    ReviewCase:
      type: object
      required: [id, case_no, case_type, responsible_team, reason_code, status, detail, allowed_actions, version, created_at]
      properties:
        id: { $ref: '#/components/schemas/Identifier' }
        case_no: { type: string }
        case_type: { type: string, minLength: 1, maxLength: 64 }
        responsible_team: { type: string }
        reason_code: { type: string }
        status: { type: string, enum: [OPEN, RESOLVED, DISMISSED] }
        order_id: { $ref: '#/components/schemas/Identifier' }
        order_line_id: { $ref: '#/components/schemas/Identifier' }
        subject_type: { type: string }
        subject_id: { $ref: '#/components/schemas/Identifier' }
        detail: { type: object, additionalProperties: true }
        suggestions:
          type: array
          items: { type: object, additionalProperties: true }
        allowed_actions:
          type: array
          items:
            type: string
            enum: [RESOLVE_CUSTOMER, RESOLVE_SKU, RESOLVE_CARRIER, RESOLVE_IMPORT_DATA, COMPLETE_SOURCE_FOLLOWUP, CONFIRM_ORDER_DRAFT, REJECT_ORDER_DRAFT, CONFIRM_TRACKING_DRAFT, OPEN_SKU_MAPPING, RERUN_JD_SKU_MAPPING_CHECK, RERUN_JD_STOCK_CHECK, REINTERPRET, REJECT]
        resolution: { type: object, additionalProperties: true, nullable: true }
        resolved_by: { type: string, nullable: true }
        resolved_at: { type: string, format: date-time, nullable: true }
        version: { type: integer, minimum: 0 }
        created_at: { type: string, format: date-time }
...